refactor: condense to single RPC, refactor to use AccountValue import
