+++ ironfish/src/rpc/routes/wallet/index.ts
@@ -16,8 +16,7 @@ export * from './getPublicKey'
 export * from './getStatus'
 export * from './getTransaction'
 export * from './getTransactions'
-export * from './importSpendAccount'
-export * from './importViewAccount'
+export * from './importAccount'
 export * from './mintAsset'
 export * from './postTransaction'
 export * from './removeAccount'
@@ -26,4 +25,3 @@ export * from './sendTransaction'
 export * from './sendTransaction'
 export * from './useAccount'
 export * from './createTransaction'
-export { ImportResponse } from './utils'
